+++ src/controller/recordController.ts
@@ -1,4 +1,5 @@
 import { Request, Response, NextFunction } from 'express';
+import { validationResult } from 'express-validator';
 import { rm, sc } from '../constants';
 import { fail, success } from '../constants/response';
 import recordService from '../service/recordService';
@@ -51,6 +52,13 @@ const deleteRecord = async (
   next: NextFunction
 ) => {
   try {
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+      return res
+        .status(sc.BAD_REQUEST)
+        .send(fail(sc.BAD_REQUEST, rm.BAD_REQUEST));
+    }
+
     const { recordId } = req.params;
 
     await recordService.deleteRecord(+recordId);
